add WaybackTweets class default values, update cli, docs
